Move logger initialization into separate module
