+++ src/cli.js
@@ -3,14 +3,11 @@
 import { ArgumentParser } from "argparse";
 import fs from "fs";
 import getPackageVersion from "@jsbits/get-package-version";
+import { initializeLog } from "./log.js";
 import { parse } from "./index.js";
 import path from "path";
-import winston from "winston";
 
-const { createLogger, format, transports } = winston;
-const { combine, timestamp, label, printf } = format;
 const version = getPackageVersion();
-
 const parser = new ArgumentParser({
   description: "Utility package for reading .nsipro files",
 });
@@ -33,28 +30,7 @@ parser.add_argument("path", {
 });
 
 let args = parser.parse_args();
-
-const loggerFormat = printf(({ level, message, label, timestamp }) => {
-  return `${timestamp} [${label}] ${level.toUpperCase()}: ${message}`;
-});
-
-const logger = createLogger({
-  level: "info",
-  format: combine(label({ label: "cli.js" }), timestamp(), loggerFormat),
-  transports: [
-    new transports.Console({
-      level: args.verbose ? "debug" : "info",
-    }),
-    new transports.File({
-      filename: "nsipro-parser.err.log",
-      level: "error",
-    }),
-    new transports.File({
-      filename: "nsipro-parser.log",
-      level: "info",
-    }),
-  ],
-});
+const logger = initializeLog(args);
 logger.info(args);
 
 const getFilesByExtension = (filepath, extension) => {
